CocoaPod bazel integration system
Summary:
This patch integrates a bazel build file compiler and repository initialization
system.

It is based on the existing logic to convert pods to bazel targets, as outlined
in the document:
https://paper.dropbox.com/doc/iOS-Bazel-External-Dependency-Integration-and-Compilation-jfdBVtuKxBk3JBjxPNkpj

The user integrates cocoapod dependencies ( and their dependencies ) into bazel
with the `pod_repository` macro.

Compiler

The compiler accepts a JSON pod spec and outputs a bazel build file. It contains
a JSON representation of the pod spec, a higher level representation of the
models ( Only ObjcLibrary for now) , and a compiler which outputs build files.

There are several differenences between code designed to work with cocoapods,
and code that bazel can compiler. Non obvious differences are:
- glob syntax.
- directory structure ( pods uses system style header directories ).
- target structure ( pods relies heavy on features in Xcode build ).
- files and dirs can have spaces in many places ( Bazel doesn't support this ).
  And more. See the linked doc for a detailed explanation.

It includes the following binaries built for OSX:

Repo tools: initialize a pod
repository to conform to the conventions of a bazel repo

PodSpecToBUILD: compile a build file from a JSON podspec

I included some notes on overall thoughts and development as well.

Test coverage:

Logic in the compiler is covered in unit tests

Build file compilation is covered in a snapshotting like test system
where we compare current outputs to known outputs in GoldMaster/