Recursive Podspec style glob
Summary:
This patch introduces a glob which is more aligned with how Pods glob works.  In
summary, it glob along paths, and includes files in each path.

Example: in the pattern Some/**/*.h Files in Some/*.h are expected to be
included

Example: in the pattern Some/**/* Files in Some/**/Foo/Bar are expected to be
included