External Dependency Smoke Tests
Summary:
This patch adds unit tests for integrating third party dependencies during the
coapod migration.  As an intermediate step of migration, we will not be able to
add cocoapods to the app until we are fully migrated. Without having full
integration into the build, there is no way to ensure that dependencies are
actually working.

It adds bazels targets: PINDependencies and PINDependenciesTests to support
testing.

A test in PINDependenciesTests is smoke test that ensures that we have:
- correctly setup a build file
- can find headers to include
- have loaded the code correctly

For now, all of the dependencies are tested in PINDependenciesTests.m.

I added TODOs for dependencies that would not work.