Command API and manual prefix fix
Summary:
This patch improves the command API which allows the running of arbitrary
commands selected by the user.

We don't blindly run "prepare_command" for a few reasons right now, so it may be
manually entered here. Examples:
- moving files
- doctoring up files with sed
- running a protobuf compiler

The user can specify any order for commands, which may be before or after repo
tools, which is also optional

Additionally, this patch implements a custom unzip and "strip prefix" feature.
Bazel is incompatible with a handful of our pods and we will move away from
manually specifying this path. Bazel's strip prefix implementation does not work
with multiple directories specified in the command either.