Implicit Pod Prefix Headers
Summary:
This patch adds support for implicit prefix headers.

There are basically 2 types of PCH's.

1) An "explicit" PCH specified in the podspec.

2) An implicit PCH that was not specified in the PodSpec, but it is required to
compile.

Case 1 is pretty straight forward, since we can just dump this into the BUILD
file.

Case 2 is not as easy and will require some logic. We can approach this with a
skylark extension that finds a PCH or returns none if there isn't one.

Implementation:

This patch adds the ability to include bazel extensions under an external
repository.

TODO:
- If a PodSpec includes a prefix header than use that instead
- remove the "paths" argument