CocoaPod Build Setting API and More
Summary:
This patch allows user configurable target attributes for Pod Deps.

This solves the following issues:
- Adding copts at the target level

    For example adding copts -foo and -bar to Bazel target Some:
    "Some.copts += -foo, -bar"

- Adding copts at the global level

   For example turning off all warnings in Pods
   and soon turning on modules

- Development Options
    A dev may need to set compiler flags during development to enable various
    features.

    Foe example, we use ASDK Flags (`AS_SAVE_EVENT_BACKTRACES` or
    `AS_ENABLE_TIPS` `AS_EVENTLOG_ENABLE`) to debug their code.

Precedence:
- Our own custom global_copts override a per pod copt
- Keypath operations override copts

API Implementation:

Input:
- The user specifies key values in the new_pod_repository macro
- We set sensible global defaults within workspace.bzl

Implementation:

All of these features leverage the same infrastructure and pipeline which include
an options parser and a type "safe" KVC API.

A new transform pass is added to apply these options.

Clang interface: This patch relies on the semantics of clang options, which are
defined in the LLVM Option parser.
https://clang.llvm.org/doxygen/classclang_1_1tooling_1_1CommonOptionsParser.html

Additionally:
- added the ability to propagate `trace` through the entire
program, which is an incredibly powerful debugging aid since we can see all of
the trace without rebuilding. It is not on by default, as this will slow down
the build!
- add radarkit as a test case