Recursive globs for alphanum ending patterns
Summary:
This patch adds support for directories ending in alphanumeric characters. For
example "Sources/Classes" or "security".

It improves the pattern factory algorithm to check for this case. It also
enforces an invariant to prevent adding this to patterns that aren't matching.

This ability is needed for a handful of pods, we've got atleast 2 in our repo.
Additionally, I added a test case where I discovered this issue,
UICollectionViewLeftAlignedLayout.