Pod IPA Processor and Validation Host
Summary:
We need to fix up bundle names.

ObjcBundleLibrary doesn't have the ability to use custom bundle names, so fix
them up before assembling the final IPA.

This is a workaround until the bazel team completes:
https://github.com/bazelbuild/rules_apple/issues/11

This will generalize to all bundling and reuses as much code as possible. I
mentioned other ways to implement this workaround in the github issue, and this
seems to be the cleanest.

Additionally, include a validation host to validation pod things. I used this in
automation programs to discover variances between the final compiled IPA and
will commit the Xcode version if we think there is value in future revisions.

We can delete this after we're full bazel

Test Plan:
Build the ValidationHost and confirm Stripe.bundle is in the output
```
tools/bazelwrapper build Platform/PINValidationHost:ios-app
unzip bazel-bin/Platform/PINValidationHost/ios-app.ipa
```