RepoTools shell magic
Summary:
We depend on the CocoaPods installation in the current environemnt. We can't
hardcore the binary because it may be running under a ruby environemnt, brew, or
some other means.

Today, this means, whatever the env Xcode is running.

The original code used the login shell to replicate the users environemnt to
determine the CocoaPods program.

This of course, is wrong ( and only works for boring shells )!

This patch adds a new `API`: shellOut(_ script: String)

Which allows executing shell scripts with the environment's current path.
Scripts are ran with `/bin/sh`.

Then, it uses the new API to find the Pods bin.