[Pods] Configurable deps & Texture.conditionals
Summary:
Texture predicates PINRemoteImage & IGListKit abilities on the availability
of includes using the `__has_include` feature in clang. Make these includes
available within the compilation sandbox.

It adds a new UserConfigurable attribute to ObjcLibrary for deps, and then uses
that to enable PINRemoteImage and IGListKit features in Texture.

Add an integration test for this feature in the build system to make sure it
works end to end.

Alternatives:
It could be benenfical to partition out the source code of Texture into
subspecs: a subspec which explicitly depends on PINRemoteImage, and a subspec
which depends on IGListKit. This is how PINRemoteImage and other pods work.

While this may be useful, it is out of scope of this project. We also want
PodSpecToBUILD to be a robust as possible, and handle code which we cannot
refactor.