[PodSpecToBUILD] Default is_dynamic to 0
Summary:
Cocoapods provides the API to link universal binarys as static libraries or
frameworks. So most Pods need to support both.

For Pinterest, we statically link, so default to this.

In the future, we can provide an API if needed.

Remove Framework stripping in the IPA processor and use dynamic for the new
Foursquare stuff, which is setup to be dynamically linked. This is a workaround
for now, and Foursquare will provide a static library.