Clang Modularization Part1
Summary:
Clang modules bring a new level of compilation performance to the table. It
improves the way that headers are utilized by the compiler. It reducing header
parsing and looking up parsed headers in a more efficient data structure.

A lot of Objective-C headers in the wild conform to the module standard: code,
which headers have been formed with care ( PINCore already works and many Pods
too ). While this only brings a small subset of our headers into modules, there
is a good boost already.

For the technical design of modules consult the clang doc
https://clang.llvm.org/docs/PCHInternals.html and source code
https://clang.llvm.org

PodSpecToBUILD

The module map generation algorithm was designed so that if a Pod has modular
headers, then it should be able to be treated as a single module. It puts all of
the `pod_support` headers into a modulemap with the name of the `bazelPodModule`

- By default, all pods are assumed to be modularized. The user specifies
  otherwise in the Workspace.
- Modules based on pod_support - other headers are not exposed ( by default )
- Codegen for module map and propagation of this map
- Use hardlinks in Pods so we can read these from the compiler
- Adding header visibility options: by default the API for pods is only in
  `pod_support`. This does not work for all cases.