Simplify command API usage
Summary:
This patch brings a few new tricks to the command API.

First, it moves it to a shell script. The dictionary was an atrocity.

Next, it removes `build_file_content`, because we should codegen a BUILD file
for all cocoapods.  If that isn't the case, then it is possible to put one in
the script and cat it into a file.

Finally, it moves curling of podspec_urls to a structured API, podspec_url. This
isn't doing anything different yet, but atleast we have a better API for the use
case.

It adds better documentation, and moves to a python style of documentation.
Eventually, this will be duplicated into a website.