Move from `external` and `repository_rule`
The original design of Pods was as idomatic `Bazel` as possible, hence why the
design used a `repository_rule`.
design used a `repository_rule`.

This is problematic for the following reasons:
- Files come and go which throw off Xcode for our users.
- Invalidation schema of `repository_rule` is not ideal.
- Doesn't work with Tulsi ( see all the hacks I had to do ).
- Users did not the cost of running this rule and downloading files, which was
  the original point of `new_pod_repository`.

- Files come and go which throw off Xcode for our users.
- Invalidation schema of `repository_rule` is not ideal.
- Doesn't work with Tulsi ( see all the hacks I had to do ).
- Users did not the cost of running this rule and downloading files, which was
  the original point of `new_pod_repository`.

Simply put, our users are paying a cost constantly when our dependencies change
and the system is unstable.

Now, once a user updates a pod, they run `update_pods.py` ( or `make pods` ).

This patch reworks PodToBUILD so we can run RepoTools fetch and init pods in Vendor/PodName.

Use '//Vendor' as the `Pods` directory, which replaces use of @ and external.

Invalidation in `update_pods.py` is based on the `.pod-version` file.
Invalidation checks all the inputs to `new_pod_repository`.

Implement `Pods.WORKSPACE`, a custom workspace variant for Pods.

The Pods.WORKSPACE, is eval'd by a python program, which has a similar effect.

User Facing API:

The user simply runs `make pods` which runs the script

Follow ups:

- The RepoTools binary and `update_pods.py` script are not currently factored
  in.
- Deletion is not happing in `update_pods.py`
- More docs.