Operator documentation and cleanup
This commit adds comments to undocumented operators, removes a
deprecated prelude operator, and deprecates the other operator that
isn't on an Apple keyboard.

What is a `Lens`, and why do we use it in? The higher level ideas are
explained in a linked blog post.

Ran `make goldmaster` to illustrate that it's still emitting the same
output.