Mixed language foundation and support
This commit adds the ability to split out sources based on the contents
of the glob. This solves for Bazel not supporting multiple languages in
the same library, for C++ or Swift.

`ObjcLibrary` can be used to compile either C++ Like or Objc Like
source.  During `BuildFile` creation, additional C++ libs are
conditionally added.  The same pattern will be used for Swift.

Additionally, break out source file rule utilities into a file.

As proof that this works, we can successfully compile texture, and now
there is a build test for the minimal example of compiling Texture.