Stabilize and assert stability of compiler output
This commit stabilizes the output of skylark code compilation. In the
compiler, it adds sorting to types that get out of order and the
ordering of such types doesn't sorts types that may be sorted:
dictionaries.

Arrays are stabilized at `toSkylark()` when possible. In general, many
outputs of PodToBUILD are already stable, due to the way that we
construct `SkylarkConvertibles`. At this current juncture, it requires
all inputs to be stable other than dictionaries.

As part of integration tests, we run multiple times.

Intention is that this Fixes #41