BuildFileContext / fix fragile header logic (#57)
Previously, we introspected label names to determine if they would emit
headers. This pattern is brittle and doesn't scale. Instead, expose all
BazelTargets during skylark generation of a BuildFile.

Originally it seemed possible to topologically sort all of the deps and
_allocate_ in reverse order. This pattern will won't work, because
transforms/simplifications can change the original inputs. Additionally.
it is less code and simpler to have access to the *entire* graph at code
gen time.