Binary release of PodToBUILD / Bazel run support (#60)
This PR adds the ability to perform release builds of `PodToBUILD`
through githubs command line automation.

Usage:
A user can point rules_pods at a release, and utilize the
`repository_rule`.

http_archive(
    name = "rules_pods",
    urls = ["https://github.com/pinterest/PodToBUILD/releases/download/__RELEASE__/PodToBUILD.zip"],
)

Additionally:

Add the ability to `bazel run` for vendorizing pods.

It creates a bazel target for `update_pods`. If the user is running a
release package, then the `Makefile` noops the build, otherwise, source
builds run with Bazel as expected. For this method, the user can add
rules_pods to the workspace in either a release `or` a source package.