Docs for bazel run, FAQ, repository_rule, installs (#61)
This PR defaults the docs to the `repository_rule` as that is the
idiomatic way to install pods in Bazel. It adds notes about using
`repository_rule`s, why `PodToBUILD` supports out of band dependency
installation, and how to `bazel run` the pod updates.

Generally, it isn't ideal to do source builds of programs like
`PodToBUILD` on a users machine for performance and build environment
coupling. This PR defaults documentation to point at binary releases.

Fixes #36