Update Apple/Swift/Bazel versions to allow building with Xcode 10.2
The goal with this is to update the dependencies to include  This updates the `rules_apple` dependency to latest in order to pull in https://github.com/bazelbuild/rules_swift/commit/19db162c614af9c9db354e14d68965c8eb58e348 and allow compiling with Xcode 10.2. This also switches to using the version of `rules_swift` included by `rules_apple`, rather than manually specifying it.

Finally, this upgrades the version of Bazel being used to be compatible with the newer versions of `rules_apple` and `rules_swift`.

This seems like it partially addresses https://github.com/pinterest/PodToBUILD/issues/74, but that also mentions other tasks like updating other projects and CI.