Use fresh temporary directories
The current implementation of `tmpdir` just returns the root of the
user's global temp dir (as returned by `NSTemporaryDirectory()`). This
changes that to create a unique subdirectory within it. This means that
each invocation of `tmpdir` will now return a new, empty directory.

This helps avoid issues with output from one repository affecting
another repository. In particular when unzipping an archive from GitHub
I was seeing issues where `unzip` was asking whether `.gitignore` should
be overwritten and immediately failing with an EOF. This should provide
better isolation between repositories.

While I'm here, this also changes a rouge tab to spaces in a Python file.