No-op on non-macOS platforms
PodToBUILD currently works by distributing a Darwin binary, which won't work on non-macOS platforms (like Linux). This is mainly an issue in non-vendored mode in a Bazel workspace shared with other, non-Apple targets. This change makes it easier for those setups by allowing the repositories to at least be loaded, even if they don't have valid content on those platforms.

In particular, this allows running `bazel query //...` on linux in a Bazel workspace that uses `new_pod_repository`.