Deterministically handle case conflicts in pod_support dir (#85)
This adds logic to de-duplicate case conflicting paths in the `pod_support` dir. This comes up in cases like the Folly pod, where the pod name is `Folly` (and therefore we try to put a copy of the headers under a `Folly` custom header search path directory) but the pod has `module_name = 'folly'` (which causes a similar set of headers under the `folly` directory). These two paths are the same on macOS due to the filesystem's case insensitivity, and which one wins is based on the ordering of the set that contains the directories, which is not well defined or deterministic across machines.

This change makes handling of this deterministic so we'll always get the same casing of these header directories, which aids in remote caching.