Make pod_support headers more deterministic (#88)
When we populate the `pod_support` directories, we put all the headers (or rather symlinks to them) in the same directory together. This means that if there are multiple headers that share a basename, only one can end up in a given directory. Previously which one was determined by the result of glob and set iteration ordering, both of which are non-deterministic (differ per run and/or machine). This sorts those headers before symlinking to guarantee that the same symlinks end up "winning out" and being used on all machines. This allows remote cache sharing/hits between machines.