Support nested pods, or pods inside of a pod (#120)
* Support nested pods, or pods inside of a pod

Adds the ability to handle "child pods" within a pod. In cocoapods, its
possible to provide a path within another pod, and this doesn't map to
any notion we have. Simply put, it generates a BUILD file in the parent
build, and aliases the Vendor/__POD__ to the parent directory.

Additionally, it has a first pass at handling adjacent, child or top
level pods. When there are 2 adjacent podspecs, cocoapods effectively
merges these into a single file. The first pass at this ability is
achieved by concatting build files together. Some algorithms rely on
having all the data structures around, so we'll need to update it to
merge build files at a higher level in the stack, and this is much more
complex.

There's a few cleanups around `EmptyBuildOptions` as it was tedious to
keep updating new values everywhere, and we might not need the protocol
either. I think this was used as a legacy feature. Will submit that as a
followup