First pass at Podfile.lock WORKSPACE generator (#119)
Add the ability to generates a transitive workspace from a Podfile.
I'm using this command to generate a Pods.WORKSPACE for transitive
dependencies, specifically with react. It emits an output that can be
used to paste into an existing workspace, or perhaps automatically
import. I'm investigating at ways to call this command automatically

Additionally, add Yams. We'll have to default to Bazel and remove SPM as
a separate PR. I needed to, version rules_swift for yams. Perhaps we
can bump to latest as the bug we have here is fixed.

Generate new_pod_repository entries from a lockfile. I expect to
automatically generate a dummy lockfile, and then import this some way

For now, it moves the build-test command out of Bazel. Running it
recursivly is very difficult to debug and had little advantages.