Implement cocoapods source file logic (#130)
* Implement cocoapods source file logic

Previously, the source file logic is close but doesn't reproduce what cocopods
does identically. For some pods, like FBSDKCoreKit, this is highly problematic.

This PR implements identical logic to what cocoapods has as of writing:
https://github.com/CocoaPods/CocoaPods/blob/master/lib/cocoapods/sandbox/file_accessor.rb

To handle arc sources it implements ruby's union operator with Globs

There's a few cleanups in this PR that make it possibl to implement the logic.
Mainly, it flips Glob<AttrSet> to AttrSet<Glob>. This enables both a recursive
glob, and also arrays of globs.

This PR also ditches the lenses which were used primarily to implement the
previous version of this. It leverages Swift 5's keypath operator instead.