Rework module maps (#149)
* Rework module maps

This PR fixes issues with module maps. First, it implements public
headers per the podspec. Module maps are normalized into a single
instance per spec. This is achieved through the
`objc_library`.module_map attribute, there is virtually no way around
using this with swift.

I've added a couple snapshot tests of the output and we'll have more
build tests for this with swift soon.