Simplify globs, simplfy rendering
This PR simplifies GlobNodes upon allocation reduce the complexity of
the glob at runtime algorithms and visually.

Logically, the output is the same, but it cleans up build files where
possible, especially in the case of combining subspec deps, arc, and
exclusions.

Additionally, it adds a few cleanups and removed the default argument
for `exclude_directories` as it's unneeded

This was pointed out in
https://github.com/pinterest/PodToBUILD/pull/153
and submitted as a follow up for simplicity