Fix issues with swift module API
1. in `swift_libray` module_name is incorrect

2.
If there's a module on the include path, clang will implicitly pick that up.

Also, it fixes the issue with cococapods:
```
[!] The following Swift pods cannot yet be integrated as static libraries:

The Swift pod  depends upon , which does not define modules. To opt into those targets generating module maps (which is necessary to import them from Swift when building as static libraries), you may set goldmaster  globally in your Podfile, or specify  for particular dependencies.
```