First pass at mixed module swift support (#153)
* First pass at mixed module swift support

This PR adds mixed module swift support with the constraints that Xcode
and the PodSpec supports. An objc library may depend on a swiftmodule
and the swift module implicitly imports the public objc module.
Currently, all sources are propagated to a single `swift_library` to
produce a single `swiftc` invocation. The swift interface header is
included in a public module map, while the public headers are exposed
to the `swiftc` invocation through a private module map

Additionally, it cleans up the rule for module map to be simpler and
more easily used in this context. The `BazelTarget` for this rule
additionally passes named arguments to improve readability.

Note: because we reduce sources into a single swift library, it must
consider subspecs that are included, otherwise, it would pull in sources
from the entire podspec. Currently it uses `Default` subspecs. As a
followup, we need to add an option to `new_pod_repository` to select
subspecs. Additionally, I'll follow up to fix the ad-hoc targets in
`ObjcLibrary`

Additionally, fix issues with swift module API

- in `swift_library` module_name is now correct

- If there's a module on the include path, we generate a modules. It ignores the default module map.

This also works great and gets around the issue with regular cococapods and Xcode:
```
[!] The following Swift pods cannot yet be integrated as static libraries:

The Swift pod  depends upon , which does not define modules. To opt into those targets generating module maps (which is necessary to import them from Swift when building as static libraries), you may set goldmaster  globally in your Podfile, or specify  for particular dependencies.
```