Build swift library with umbrella header import
This reproduces identically what CocoaPods + Xcode is doing. There's a
few hidden side effects of not building pods as mixed modules which is
apparent in a few pods. The main issue I've found so far, is implicit
importing of umbrella header contents that doesn't happen.

This is not 100% the most effective approach but reproduces identically
what CocoaPods + Xcode is doing. In the future, we should consider
reducing module map building, but it will need more testing to deviate
from the default behavior. For now, this atleast gets it working.