Add a build test for top 50 swift pods (#159)
* Add a build test for top 50 swift pods

The swift support implemented should build almost every cocoapod. I've
pulled the top 50 swift repos from github and generated a test case from
it. It disabled pods that don't work on Swift 5, as they won't work on
Xcode

I've also checked in a simple python script to gen this test case and
transitive deps. For now, it performs a search and then looks up the pod
from the cocoapods search cache. We can probably improve this code ( and
the workspace generator ) by using the cocoapods spec repo. Mainly
checking this in to document the current algorithm to perfrom the
search.