Allow the replacement of execpath
In order to use `expand_location` we need to have a dep on the input.
https://docs.bazel.build/versions/master/skylark/lib/ctx.html#expand_location

Private module maps used in swift library that were previously not
`deps`, and they are specified now specified as header, so they remain
private. If we fix `swiftc_inputs` to be expandable in Bazel this would
be unnecessary